update docusaurus.config.js
updated file to point to welcome.md instead of intro.md after the file rename

this solved the error that caused the last two reverts to find the bug. It was point to a non existant file, so serving to localhost rendered and error that just displayed the file names.